R&D 회차 5 — 하드코딩 퀴즈를 콘텐츠로, 채점 없던 문제 수정
index.html 에 박혀 있던 퀴즈를 content/concept/q01.md 로 옮겼다.
콘텐츠 파일에 없는 내용이 배포되던 문제이자, splice(2,...) 로 위치
2에 끼어들어 목차의 "01 수열의 극한" 절을 둘로 가르던 원인이다.
이제 01절(5개) → 확인 문제 → 02 급수 로 각 머리글이 한 번씩만 나온다.

build.py 의 order 를 float 으로 바꿔 5.5 같은 값을 허용했다. 개념은
파일번호=order 관례를 그대로 두고, 사이에 끼우는 화면만 소수로 넣기
위해서다(하드코딩 퀴즈도 원래 order 2.5 였다).

브라우저로 확인하다 새 결함을 찾았다: answer 필드를 build.py 가 읽고
미니 에디터가 입력까지 받는데 화면에서는 쓰이지 않았다. 보기를 고르면
선택 표시와 효과음만 날 뿐 맞았는지 알려주지 않아, 앱의 유일한 확인
문제가 확인을 못 하고 있었다. 고른 답이 틀리면 정답을 함께 드러내도록
채점을 붙였고 오답·정답 양쪽을 실제로 눌러 확인했다.

+++ mathuit/build.py
@@ -142,6 +142,9 @@ def parse_tip(path):
                  'lines': lines, 'lab': {'x': lx, 'y': ly, 'a': at, 't': syms(g.get('label',''))},
                  'steps': steps, 'rel': meta.get('related', []), 'nograph': not bool(gm)}
 
+# order 는 정수가 아니라 실수다. 개념은 파일번호와 같은 정수를 쓰지만(c07 → 7),
+# 퀴즈처럼 개념 사이에 끼우는 화면은 5.5 같은 값으로 넣어야 앞뒤 개념의 번호를
+# 전부 밀지 않아도 된다. 파일번호=order 관례를 깨지 않기 위한 선택이다.
 def parse_concept(path):
     f = os.path.basename(path)
     meta, body = front(open(path, encoding='utf-8').read(), f)
@@ -151,7 +154,7 @@ def parse_concept(path):
         options = []
         for i in range(1, 6):
             options.append(meta.get('option%d' % i, ''))
-        return {'id': f[:-3], 'order': int(meta.get('order', 999)), 'no': meta['no'], 'unit': meta['unit'],
+        return {'id': f[:-3], 'order': float(meta.get('order', 999)), 'no': meta['no'], 'unit': meta['unit'],
                 'sec': meta['sec'], 'title': meta['title'], 'kind': 'quiz',
                 'question': meta.get('question', ''), 'problemEq': meta.get('problemEq', ''),
                 'options': options, 'answer': int(meta.get('answer', 1)),
@@ -176,7 +179,7 @@ def parse_concept(path):
     # 이 관계는 콘텐츠의 일부다 — check.py 가 존재·순서를 검사한다.
     pre = meta.get('prereq', [])
     if isinstance(pre, str): pre = [x.strip() for x in pre.split(',') if x.strip()]
-    return {'id': f[:-3], 'order': int(meta.get('order', 999)), 'no': meta['no'], 'unit': meta['unit'],
+    return {'id': f[:-3], 'order': float(meta.get('order', 999)), 'no': meta['no'], 'unit': meta['unit'],
             'sec': meta['sec'], 'title': meta['title'], 'pre': pre,
             'p1': prose(before), 'p2': prose(after), 'eqL': eqL, 'eqR': eqR}
